+++ src/lib.rs
@@ -117,6 +117,7 @@ pub mod locale_model;
 pub mod lorem;
 pub mod lsp_core;
 pub mod nerd_font_picker;
+pub mod org;
 pub mod outline_panel;
 pub mod pomodoro_tool;
 pub mod right_dock;